CColorStrategyPool::CheckColorMODE の実装をヘッダに置いて inline 化
