+++ sakura_core/view/colors/CColorStrategy.h
@@ -170,67 +170,5 @@ protected:
 	const STypeConfig* m_pTypeData;
 };
 
-#include "util/design_template.h"
-#include <vector>
-class CColor_LineComment;
-class CColor_BlockComment;
-class CColor_BlockComment;
-class CColor_SingleQuote;
-class CColor_DoubleQuote;
-class CColor_Heredoc;
-
-class CColorStrategyPool : public TSingleton<CColorStrategyPool>{
-	friend class TSingleton<CColorStrategyPool>;
-	CColorStrategyPool();
-	virtual ~CColorStrategyPool();
-
-public:
-
-	//取得
-	CColorStrategy*	GetStrategy(int nIndex) const{ return m_vStrategiesDisp[nIndex]; }
-	int				GetStrategyCount() const{ return (int)m_vStrategiesDisp.size(); }
-	CColorStrategy*	GetStrategyByColor(EColorIndexType eColor) const;
-
-	//特定取得
-	CColor_Found*   GetFoundStrategy() const{ return m_pcFoundStrategy; }
-	CColor_Select*  GetSelectStrategy() const{ return m_pcSelectStrategy; }
-
-	//イベント
-	void NotifyOnStartScanLogic();
-
-	/*
-	|| 色分け
-	*/
-	//@@@ 2002.09.22 YAZAKI
-	// 2005.11.21 Moca 引用符の色分け情報を引数から除去
-	void CheckColorMODE( CColorStrategy** ppcColorStrategy, int nPos, const CStringRef& cLineStr );
-	bool IsSkipBeforeLayout();	// レイアウトが行頭からチェックしなくていいか判定
-
-	//設定変更
-	void OnChangeSetting(void);
-
-	//ビューの設定・取得
-	CEditView* GetCurrentView(void) const{ return m_pcView; }
-	void SetCurrentView(CEditView* pcView) { m_pcView = pcView; }
-
-private:
-	std::vector<CColorStrategy*>	m_vStrategies;
-	std::vector<CColorStrategy*>	m_vStrategiesDisp;	//!< 色分け表示対象
-	CColor_Found*					m_pcFoundStrategy;
-	CColor_Select*					m_pcSelectStrategy;
-
-	CColor_LineComment*				m_pcLineComment;
-	CColor_BlockComment*			m_pcBlockComment1;
-	CColor_BlockComment*			m_pcBlockComment2;
-	CColor_SingleQuote*				m_pcSingleQuote;
-	CColor_DoubleQuote*				m_pcDoubleQuote;
-	CColor_Heredoc*					m_pcHeredoc;
-
-	CEditView*						m_pcView;
-
-	bool	m_bSkipBeforeLayoutGeneral;
-	bool	m_bSkipBeforeLayoutFound;
-};
-
 #endif /* SAKURA_CCOLORSTRATEGY_BC7B5956_A0AF_4C9C_9C0E_07FE658028AC9_H_ */
 /*[EOF]*/
